made game speed up earlier added pause continuation countdown fixed popup behavior when turning off the screen

+++ src/com/mammothGames/wake1/gameEngine/engine_entity.java
@@ -34,7 +34,7 @@ public class engine_entity {
 	}
 	
 	/////////////////ALARMS/////////////////
-	public double[] alarm = new double[13];
+	public float[] alarm = new float[13];
 	protected void sys_initAlarms(){
 		alarm[0] = -1;
 		alarm[1] = -1;
